+++ Game.Server/Packets/Client/HotSpringRoomEnterViewHandler.cs
@@ -12,36 +12,27 @@ namespace Game.Server.Packets.Client
     {
         public int HandlePacket(GameClient client, GSPacketIn packet)
         {
-            //if (client.Player.CurrentHotSpringRoom != null)
-            //{
-            //    client.Player.CurrentHotSpringRoom.ProcessData(client.Player, packet);
-            //}
-            GSPacketIn pkg = new GSPacketIn((byte)ePackageType.HOTSPRING_ROOM_PLAYER_ADD);
+            GSPacketIn pkg = new GSPacketIn((short)ePackageType.HOTSPRING_ROOM_PLAYER_ADD);
+            
             PlayerInfo player = client.Player.PlayerCharacter;
+
             pkg.WriteInt(player.ID);
             pkg.WriteInt(player.Grade);
             pkg.WriteInt(player.Hide);
-
             pkg.WriteInt(player.Repute);
             pkg.WriteString(player.NickName);
             pkg.WriteBoolean(true);
             pkg.WriteInt(5);
-  
             pkg.WriteBoolean(player.Sex);
             pkg.WriteString(player.Style);
             pkg.WriteString(player.Colors);
             pkg.WriteString(player.Skin);
             pkg.WriteInt(300);
             pkg.WriteInt(400);
-            //var _loc_6:* = new Point(_loc_2.readInt(), _loc_2.readInt());
             pkg.WriteInt(player.FightPower);
             pkg.WriteInt(player.Win);
             pkg.WriteInt(player.Total);
             pkg.WriteInt(1);
-            //_loc_5.FightPower = _loc_2.readInt();
-            //_loc_5.WinCount = _loc_2.readInt();
-            //_loc_5.TotalCount = _loc_2.readInt();
-            //_loc_4.playerDirection = _loc_2.readInt();
             client.SendTCP(pkg);
             return 0;
         }
